Architecture rule breach fixes
