+++ cms/metrics_interface/interface.py
@@ -5,8 +5,7 @@ from django.db.models import Manager, QuerySet
 from metrics.data.models import core_models
 from metrics.domain.charts.line_multi_coloured.colour_scheme import RGBAColours
 from metrics.domain.charts.line_multi_coloured.properties import ChartLineTypes
-from metrics.domain.utils import GraphAxisFields
-from metrics.interfaces.charts.access import ChartTypes
+from metrics.interfaces.charts.access import ChartAxisFields, ChartTypes
 
 DEFAULT_TOPIC_MANAGER = core_models.Topic.objects
 DEFAULT_METRIC_MANAGER = core_models.Metric.objects
@@ -72,7 +71,7 @@ class MetricsAPIInterface:
     @staticmethod
     def get_chart_axis_choices() -> List[Tuple[str, str]]:
         """Gets all available axis choices for a chart as a list of 2-item tuples.
-        Note this is achieved by delegating the call to the `GraphAxisFields` enum from the Metrics API
+        Note this is achieved by delegating the call to the `ChartAxisFields` enum from the Metrics API
 
         Returns:
             List[Tuple[str, str]]: List of 2 item tuples as expected by the form blocks.
@@ -80,7 +79,7 @@ class MetricsAPIInterface:
                 [("geography", "geography"), ...]
 
         """
-        return GraphAxisFields.choices()
+        return ChartAxisFields.choices()
 
     @staticmethod
     def get_chart_line_types() -> List[Tuple[str, str]]:
