+++ metrics/interfaces/charts/access.py
@@ -1,4 +1,5 @@
 import datetime
+from enum import StrEnum
 from typing import Dict, List, Optional, Union
 
 import plotly.graph_objects
@@ -443,3 +444,14 @@ def write_figure(
     figure.write_image(file=filename, format=file_format)
 
     return filename
+
+
+class ChartAxisFields(StrEnum):
+    stratum = "stratum__name"
+    date = "dt"
+    metric = "metric_value"
+    geography = "geography__geography_type__name"
+
+    @classmethod
+    def choices(cls):
+        return tuple((field_name.name, field_name.name) for field_name in cls)
